Factor out errorHandler and related code into httputil package.
This functionality will be needed from multiple packages in the future.

Name it httputil for the lack of a better name.
