+++ httputil/doc.go
@@ -1,0 +1,2 @@
+// Package httputil is a custom HTTP framework created specifically for home.
+package httputil
