+++ notifications.go
@@ -12,6 +12,7 @@ import (
 	"github.com/google/go-github/github"
 	"github.com/gregjones/httpcache"
 	"github.com/shurcooL/home/component"
+	"github.com/shurcooL/home/httputil"
 	"github.com/shurcooL/htmlg"
 	"github.com/shurcooL/notifications"
 	"github.com/shurcooL/notifications/fs"
@@ -29,13 +30,13 @@ type notificationsAPIHandler struct {
 
 func (h notificationsAPIHandler) Count(w http.ResponseWriter, req *http.Request) error {
 	if req.Method != "GET" {
-		return MethodError{Allowed: []string{"GET"}}
+		return httputil.MethodError{Allowed: []string{"GET"}}
 	}
 	n, err := h.notifications.Count(req.Context(), nil)
 	if err != nil {
 		return err
 	}
-	return JSONResponse{n}
+	return httputil.JSONResponse{n}
 }
 
 // initNotifications creates and returns a notification service,
@@ -63,7 +64,7 @@ func initNotifications(root webdav.FileSystem, users users.Service) (notificatio
 
 	// Register HTTP API endpoint.
 	notificationsAPIHandler := notificationsAPIHandler{notifications: service}
-	http.Handle("/api/notifications/count", userMiddleware{errorHandler{notificationsAPIHandler.Count}})
+	http.Handle("/api/notifications/count", userMiddleware{httputil.ErrorHandler{notificationsAPIHandler.Count}})
 
 	// Register notifications app endpoints.
 	opt := notificationsapp.Options{
