+++ resume.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"strconv"
 
+	"github.com/shurcooL/home/httputil"
 	"github.com/shurcooL/notifications"
 	"github.com/shurcooL/reactions"
 	"github.com/shurcooL/resume"
@@ -39,9 +40,9 @@ const googleAnalytics = `<script>
 
 // resumeJSCSS contains /resume.{js,css}.
 func initResume(resumeJSCSS http.Handler, reactions reactions.Service, notifications notifications.Service, usersService users.Service) {
-	http.Handle("/resume", userMiddleware{errorHandler{func(w http.ResponseWriter, req *http.Request) error {
+	http.Handle("/resume", userMiddleware{httputil.ErrorHandler{func(w http.ResponseWriter, req *http.Request) error {
 		if req.Method != "GET" {
-			return MethodError{Allowed: []string{"GET"}}
+			return httputil.MethodError{Allowed: []string{"GET"}}
 		}
 
 		w.Header().Set("Content-Type", "text/html; charset=utf-8")
