+++ usercontent.go
@@ -1,16 +1,14 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"io"
-	"log"
 	"net/http"
 	"os"
 	pathpkg "path"
-	"strings"
 
 	"github.com/satori/go.uuid"
+	"github.com/shurcooL/home/httputil"
 	"github.com/shurcooL/users"
 	"github.com/shurcooL/webdavfs/vfsutil"
 	"golang.org/x/net/webdav"
@@ -23,7 +21,7 @@ type userContentHandler struct {
 
 func (uc userContentHandler) Upload(w http.ResponseWriter, req *http.Request) error {
 	if req.Method != "POST" {
-		return MethodError{Allowed: []string{"POST"}}
+		return httputil.MethodError{Allowed: []string{"POST"}}
 	}
 
 	type uploadResponse struct {
@@ -33,27 +31,27 @@ func (uc userContentHandler) Upload(w http.ResponseWriter, req *http.Request) er
 
 	user, err := uc.users.GetAuthenticated(req.Context())
 	if err != nil {
-		return JSONResponse{uploadResponse{Error: err.Error()}}
+		return httputil.JSONResponse{uploadResponse{Error: err.Error()}}
 	}
 	if user.ID == 0 {
-		return JSONResponse{uploadResponse{Error: os.ErrPermission.Error()}}
+		return httputil.JSONResponse{uploadResponse{Error: os.ErrPermission.Error()}}
 	}
 
 	if contentType := req.Header.Get("Content-Type"); contentType != "image/png" {
-		return JSONResponse{uploadResponse{Error: fmt.Sprintf("Content-Type %q is not supported", contentType)}}
+		return httputil.JSONResponse{uploadResponse{Error: fmt.Sprintf("Content-Type %q is not supported", contentType)}}
 	}
 
 	dir := fmt.Sprintf("/%d@%s", user.ID, user.Domain)
 	err = vfsutil.MkdirAll(uc.store, dir, 0755)
 	if err != nil {
-		return JSONResponse{uploadResponse{Error: err.Error()}}
+		return httputil.JSONResponse{uploadResponse{Error: err.Error()}}
 	}
 
 	name := uuid.NewV4().String() + ".png"
 	path := pathpkg.Join(dir, name)
 	f, err := uc.store.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0644)
 	if err != nil {
-		return JSONResponse{uploadResponse{Error: err.Error()}}
+		return httputil.JSONResponse{uploadResponse{Error: err.Error()}}
 	}
 
 	const maxSizeBytes = 10 * 1024 * 1024
@@ -62,20 +60,20 @@ func (uc userContentHandler) Upload(w http.ResponseWriter, req *http.Request) er
 	if err != nil {
 		f.Close()
 		uc.store.RemoveAll(path)
-		return JSONResponse{uploadResponse{Error: err.Error()}}
+		return httputil.JSONResponse{uploadResponse{Error: err.Error()}}
 	}
 	err = f.Close()
 	if err != nil {
 		uc.store.RemoveAll(path)
-		return JSONResponse{uploadResponse{Error: err.Error()}}
+		return httputil.JSONResponse{uploadResponse{Error: err.Error()}}
 	}
 
-	return JSONResponse{uploadResponse{URL: pathpkg.Join("/usercontent", path)}}
+	return httputil.JSONResponse{uploadResponse{URL: pathpkg.Join("/usercontent", path)}}
 }
 
 func (uc userContentHandler) Serve(w http.ResponseWriter, req *http.Request) error {
 	if req.Method != "GET" {
-		return MethodError{Allowed: []string{"GET"}}
+		return httputil.MethodError{Allowed: []string{"GET"}}
 	}
 
 	f, err := vfsutil.Open(uc.store, req.URL.Path)
@@ -91,78 +89,4 @@ func (uc userContentHandler) Serve(w http.ResponseWriter, req *http.Request) err
 	w.Header().Set("Content-Type", "image/png")
 	http.ServeContent(w, req, "", fi.ModTime(), f)
 	return nil
-}
-
-// errorHandler factors error handling out of the HTTP handler.
-type errorHandler struct {
-	handler func(w http.ResponseWriter, req *http.Request) error
-}
-
-func (h errorHandler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
-	rw := &responseWriter{ResponseWriter: w}
-	err := h.handler(rw, req)
-	switch {
-	case err != nil && rw.WroteHeader:
-		// The header has already been written, so it's too late to send
-		// a different status code. Just log the error and move on.
-		log.Println(err)
-	case IsMethodError(err):
-		w.Header().Set("Allow", strings.Join(err.(MethodError).Allowed, ", "))
-		http.Error(w, err.Error(), http.StatusMethodNotAllowed)
-	case IsRedirect(err):
-		http.Redirect(w, req, err.(Redirect).URL, http.StatusSeeOther)
-	case IsHTTPError(err):
-		http.Error(w, err.Error(), err.(HTTPError).Code)
-	case IsJSONResponse(err):
-		w.Header().Set("Content-Type", "application/json")
-		jw := json.NewEncoder(w)
-		jw.SetIndent("", "\t")
-		err := jw.Encode(err.(JSONResponse).V)
-		if err != nil {
-			log.Println("error encoding JSONResponse:", err)
-		}
-	case os.IsNotExist(err):
-		log.Println(err)
-		http.Error(w, err.Error(), http.StatusNotFound)
-	case os.IsPermission(err):
-		log.Println(err)
-		http.Error(w, err.Error(), http.StatusUnauthorized)
-	default:
-		// Do nothing.
-	case err != nil:
-		log.Println(err)
-		// TODO: Only display error details to SiteAdmin users?
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-	}
-}
-
-// responseWriter wraps a real http.ResponseWriter and captures
-// whether or not the header has been written.
-type responseWriter struct {
-	http.ResponseWriter
-
-	WroteHeader bool // Write or WriteHeader was called.
-}
-
-func (rw *responseWriter) Write(p []byte) (n int, err error) {
-	rw.WroteHeader = true
-	return rw.ResponseWriter.Write(p)
-}
-func (rw *responseWriter) WriteHeader(code int) {
-	rw.WroteHeader = true
-	rw.ResponseWriter.WriteHeader(code)
-}
-
-// MethodError is an error type used for methods that aren't allowed.
-type MethodError struct {
-	Allowed []string // Allowed methods.
-}
-
-func (m MethodError) Error() string {
-	return fmt.Sprintf("method should be %v", strings.Join(m.Allowed, " or "))
-}
-
-func IsMethodError(err error) bool {
-	_, ok := err.(MethodError)
-	return ok
 }
